+++ crates/sncast/src/starknet_commands/tx_status.rs
@@ -1,22 +1,24 @@
 use anyhow::{Context, Result};
 use clap::Args;
-use starknet::core::types::{FieldElement, TransactionStatus};
+use sncast::response::structs::TransactionStatusResponse;
+use starknet::core::types::FieldElement;
 use starknet::providers::jsonrpc::HttpTransport;
 use starknet::providers::{JsonRpcClient, Provider};
 
 #[derive(Args)]
 #[command(about = "Get the status of a transaction")]
 pub struct TxStatus {
-    /// Hash of the transaction to get the status
+    /// Hash of the transaction
     pub transaction_hash: FieldElement,
 }
 
 pub async fn tx_status(
     provider: &JsonRpcClient<HttpTransport>,
     transaction_hash: FieldElement,
-) -> Result<TransactionStatus> {
+) -> Result<TransactionStatusResponse> {
     provider
         .get_transaction_status(transaction_hash)
         .await
         .context("Failed to get transaction status")
+        .map(Into::into)
 }