+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- `tx-status` subcommand to get transaction status
+- `tx-status` subcommand to get transaction status. [Read more here](./docs/src/starknet/tx-status.md)
 
 ## [0.24.0] - 2024-05-22